app/blueprint-app/requirements.txt
Flask
azure-identity
azure-mgmt-resource
azure-mgmt-storage
azure-mgmt-containerregistry
azure-mgmt-appcontainers